assignment update

+++ src/controller/post_controller.py
@@ -1,6 +1,7 @@
 from src.configs.connections.mysql import get_mysql_connection
 from src.service.notification.notification_service import NotificationService
 from src.repository.mongodb.post import PostRepository
+from src.repository.mongodb.classroom import MongoClassroomRepository
 from fastapi import APIRouter, Depends, HTTPException, UploadFile, File, Form
 from src.configs.connections import get_mongo_connection
 from typing import List
@@ -8,6 +9,8 @@ from src.service.models.classroom import PostResponse, PostCreate, PostUpdate
 from pymongo.errors import PyMongoError
 from src.configs.connections.blob_storage import SupabaseStorage
 import os
+import uuid
+
 
 BUCKET = 'posts'
 POST_CONTROLLER = APIRouter(tags=['Post'])
@@ -23,7 +26,7 @@ current_user = {
 
 
 @POST_CONTROLLER.get("/classroom/{class_id}/post/all", response_model=List[PostResponse])
-async def get_all_posts(class_id: str, connection=Depends(get_mongo_connection)):
+async def get_all_posts(class_id: str, connection=Depends(get_mongo_connection)) -> List[PostResponse]:
     try:
         # TODO: replace with user from token
         # ensure user is logged in
@@ -31,30 +34,42 @@ async def get_all_posts(class_id: str, connection=Depends(get_mongo_connection))
             raise HTTPException(status_code=403, detail='Unauthorized. You must login before accessing this resource.')
 
         # ensure user is a participant of the class
-        repo = PostRepository(connection)
-        participants = await repo.get_participants(class_id)
+        classroom_repo = MongoClassroomRepository(connection)
+        post_repo = PostRepository(connection)
+        if not await classroom_repo.find_participant_in_class(current_user['id'], class_id):
+            raise HTTPException(status_code=403, detail='Unauthorized. You must be a participant of the class.')
 
-        posts = await repo.get_posts_in_class(class_id)
+        posts = await post_repo.get_posts_in_class(class_id)
         return [PostResponse(**post) for post in posts]
     except PyMongoError as e:
         raise HTTPException(status_code=500, detail=f"Database MongoDB error: {str(e)}")
 
 
 @POST_CONTROLLER.get("/classroom/{class_id}/post/{post_id}/detail", response_model=PostResponse)
-async def get_post_by_id(class_id: str, post_id: str, connection=Depends(get_mongo_connection)):
+async def get_post_by_id(class_id: str, post_id: str, connection=Depends(get_mongo_connection)) -> PostResponse:
     try:
-        repo = PostRepository(connection)
-        storage = SupabaseStorage()
-        post_folder = os.path.join(class_id, post_id)
+        # TODO: replace with user from token
+        # ensure user is logged in
+        if not current_user:
+            raise HTTPException(status_code=403, detail='Unauthorized. You must login before accessing this resource.')
 
-        db_post = await repo.get_post_by_id(class_id, post_id)
+        # ensure user is a participant of the class
+        classroom_repo = MongoClassroomRepository(connection)
+        if not await classroom_repo.find_participant_in_class(current_user['id'], class_id):
+            raise HTTPException(status_code=403, detail='Unauthorized. You must be a participant of the class.')
+
+        post_repo = PostRepository(connection)
+        storage = SupabaseStorage()
+
+        db_post = await post_repo.get_post_by_id(class_id, post_id)
         if db_post is None:
             raise HTTPException(status_code=404, detail="Post not found")
 
         # generate url for each attachment
+        post_folder = class_id + '/' + post_id
         urls = await storage.get_file_urls(
             bucket_name=BUCKET,
-            file_locations=[os.path.join(post_folder, filename) for filename in db_post['attachments']]
+            file_locations=[post_folder + '/' + filename for filename in db_post['attachments']]
         )
         db_post['attachments'] = urls
 
@@ -71,26 +86,34 @@ async def create_post(
         attachments: List[UploadFile] = File(None),
         connection=Depends(get_mongo_connection),
         mysql_cnx=Depends(get_mysql_connection)
-):
-    author = 'hoangkimgiap'  # temporary author, will be replaced by username from token
+) -> dict:
 
     try:
+        # TODO: replace with user from token
+        # ensure user is logged in
+        if not current_user:
+            raise HTTPException(status_code=403, detail='Unauthorized. You must login before accessing this resource.')
+
+        # ensure user is a participant of the class
+        classroom_repo = MongoClassroomRepository(connection)
+        if not await classroom_repo.find_participant_in_class(current_user['id'], class_id):
+            raise HTTPException(status_code=403, detail='Unauthorized. You must be a participant of the class.')
+
         # create post in database
         repo = PostRepository(connection)
         post_dict = {
-            "title": title if title else None,
-            "content": content if content else None,
+            "title": title,
+            "content": content,
             "attachments": [file.filename for file in attachments] if attachments else None
         }
         post_dict = {k: v for k, v in post_dict.items() if v is not None}
-        new_post = PostCreate(**post_dict, author=author)
-        newpost_id = await repo.create_post(class_id=class_id, new_post=new_post)
-        if not newpost_id:
+        newpost_id = 'post-' + str(uuid.uuid4())
+        new_post = PostCreate(**post_dict, id=newpost_id, author=current_user['username'])
+        if not await repo.create_post(class_id=class_id, new_post=new_post):
             raise HTTPException(status_code=500, detail='An unexpected error occurred. Create post failed')
 
         # upload attachments to storage
         storage = SupabaseStorage()
-        # TODO: fix for Windows
         post_folder = class_id + "/" + newpost_id
         upload_results = await storage.bulk_upload(
             bucket_name=BUCKET,
@@ -98,15 +121,17 @@ async def create_post(
             dest_folder=post_folder
         )
 
+        # create notification for students
         notification_service = NotificationService(class_id, mysql_cnx)
         notification_service.create_new_notification_for_students(
-            title=author + " has created a new post.",
+            title=current_user['username'] + " has created a new post.",
             content=title,
             direct_url=f"/c/{class_id}/p/{newpost_id}"
         )
 
         return {
             'message': 'Post created successfully',
+            'post_id': newpost_id,
             'upload_results': upload_results,
         }
     except PyMongoError as e:
@@ -122,12 +147,20 @@ async def update_post(
         additional_attachments: List[UploadFile] = File(None),
         removal_attachments: List[str] = Form(None),
         connection=Depends(get_mongo_connection)
-):
+) -> dict:
     try:
-        # update post in database
-        repo = PostRepository(connection)
-        post_folder = os.path.join(class_id, post_id)
+        # TODO: replace with user from token
+        # ensure user is logged in
+        if not current_user:
+            raise HTTPException(status_code=403, detail='Unauthorized. You must login before accessing this resource.')
 
+        # ensure user is post's author
+        repo = PostRepository(connection)
+        db_post = repo.get_post_by_id(class_id, post_id)
+        if current_user['username'] != db_post['author']:
+            raise HTTPException(status_code=403, detail='Unauthorized. You must be the author of the post.')
+
+        # update post in database
         update_data_dict = {
             "title": title if title else None,
             "content": content if content else None,
@@ -143,6 +176,7 @@ async def update_post(
             raise HTTPException(status_code=500, detail='An unexpected error occurred. Update post failed')
 
         # upload and remove attachments
+        post_folder = class_id + '/' + post_id
         storage = SupabaseStorage()
         upload_results = await storage.bulk_upload(
             bucket_name=BUCKET,
@@ -152,7 +186,7 @@ async def update_post(
 
         remove_results = await storage.remove_files(
             bucket_name=BUCKET,
-            file_locations=[os.path.join(post_folder, filename) for filename in removal_attachments]
+            file_locations=[post_folder + '/' + filename for filename in removal_attachments]
             if removal_attachments else []
         )
 
@@ -168,8 +202,19 @@ async def update_post(
 
 
 @POST_CONTROLLER.delete("/classroom/{class_id}/post/{post_id}/delete")
-async def delete_post(class_id: str, post_id: str, connection=Depends(get_mongo_connection)):
+async def delete_post(class_id: str, post_id: str, connection=Depends(get_mongo_connection)) -> dict:
     try:
+        # TODO: replace with user from token
+        # ensure user is logged in
+        if not current_user:
+            raise HTTPException(status_code=403, detail='Unauthorized. You must login before accessing this resource.')
+
+        # ensure user is post's author
+        repo = PostRepository(connection)
+        db_post = await repo.get_post_by_id(class_id, post_id)
+        if current_user['username'] != db_post['author']:
+            raise HTTPException(status_code=403, detail='Unauthorized. You must be the author of the post.')
+
         # delete post in database
         repo = PostRepository(connection)
         status = await repo.delete_post_by_id(class_id, post_id)
@@ -178,10 +223,10 @@ async def delete_post(class_id: str, post_id: str, connection=Depends(get_mongo_
 
         # delete attachments in storage
         storage = SupabaseStorage()
-        post_folder = os.path.join(class_id, post_id)
-        remove_results = await storage.remove_folder(
+        post_folder = class_id +'/' + post_id
+        remove_results = await storage.remove_files(
             bucket_name=BUCKET,
-            folder_path=post_folder
+            file_locations=[post_folder + '/' + filename for filename in db_post['attachments']]
         )
 
         return {
